// lib/api/generated/model/updateFaqCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export interface UpdateFaqCommand {
  /** @nullable */
  question?: string | null;
  /** @nullable */
  answer?: string | null;
  isActive?: boolean;
  sortOrder?: number;
}